Move LocalizedString to Sandra.UI.csproj.
