+++ Eutherion/Shared/Eutherion.Shared.csproj
@@ -39,7 +39,7 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="LinqExtensions.cs" />
-    <Compile Include="Localization\LocalizedString.cs" />
+    <Compile Include="Localization\LocalizedStringKey.cs" />
     <Compile Include="Localization\Localizer.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Text\Json\JsonBooleanLiteralSyntax.cs" />
