+++ Sandra.UI/Sandra.UI.csproj
@@ -62,6 +62,7 @@
     <Compile Include="AppTemplate\ListBoxEx.cs">
       <SubType>Component</SubType>
     </Compile>
+    <Compile Include="AppTemplate\LocalizedString.cs" />
     <Compile Include="AppTemplate\Localizers.cs" />
     <Compile Include="AppTemplate\MenuCaptionBarForm.cs">
       <SubType>Form</SubType>
